tidy dependencies
